Fix build: 4 compile errors from the interface overhaul
1) HabitsListScreen.kt: 'Unresolved reference graphicsLayer' at the import
   itself (wrong package guessed). Switched the drag-offset visual effect
   from Modifier.graphicsLayer { translationY = ... } to the more standard
   Modifier.offset { IntOffset(0, dragOffset.roundToInt()) } pattern
   (androidx.compose.foundation.layout.offset) - the idiomatic, widely-used
   approach for drag-reorder in Compose, and one I'm confident of the
   import path for.

2) HistoryScreen.kt:434: same class of bug as the earlier arrayOf(String,
   Long) issue - (if (pct <= 0) 8 else 8 + pct * 0.5).dp mixed an Int
   branch (8) with a Double branch (8 + pct * 0.5); Kotlin doesn't
   auto-widen Int to Double across if-branches, so the inferred type
   matched none of the Int/Float/Double .dp extensions. Fixed by making
   both branches explicitly Double (8.0).

3) HomeScreen.kt:68: (0..6).map { weekMonday.plusDays(it) } - 'it' here is
   Int (from the literal IntRange 0..6), but plusDays expects Long.
   Added .toLong(). Audited every other plusDays/minusDays call in the new
   code: the ones using ChronoUnit.DAYS.between(...) as a range bound are
   fine (that returns Long, so 0..thatLong is already a LongRange), and
   bare integer literals like plusDays(6) are fine too (Kotlin infers
   literal constants as Long from context) - this was the only real
   instance of the mismatch.

+++ app/src/main/java/com/atomichabits/tracker/ui/home/HomeScreen.kt
@@ -65,7 +65,7 @@ fun HomeScreen(
     // falls rather than always being the last/rightmost item.
     val weekMonday = remember(today) { today.minusDays((today.dayOfWeek.value - 1).toLong()) }
     val weekSunday = remember(weekMonday) { weekMonday.plusDays(6) }
-    val days = remember(weekMonday) { (0..6).map { weekMonday.plusDays(it) } }
+    val days = remember(weekMonday) { (0..6).map { weekMonday.plusDays(it.toLong()) } }
 
     val windowLogs by app.repository.observeLogsBetween(weekMonday, weekSunday).collectAsState(initial = emptyList())
     val scope = rememberCoroutineScope()
